requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distro-info===0.18ubuntu0.18.04.1
Django==3.0.2
django-filter==2.2.0
django-import-export==2.0.2
django-leaflet==0.26.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
feedparser==5.2.1
guacamole==0.9.2
html5lib==1.0.1
httpie==0.9.8
httplib2==0.9.2
idna==2.6
jdcal==1.4.1
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==3.1.2
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
oauthlib==2.0.6
odfpy==1.4.1
olefile==0.45.1
onboard==1.4.1
openpyxl==3.0.3
padme==1.1.1
pandas==0.25.3
PasteDeploy==2.0.1
pexpect==4.2.1
Pillow==5.1.0
plainbox==0.25
protobuf==3.0.0
psycopg2==2.7.4
psycopg2-binary==2.8.4
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
PyPDF2==1.26.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-gettext==4.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tablib==1.1.0
transaction==3.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.2
webencodings==0.5
xdiagnose==3.8.8
xhtml2pdf==0.2.4
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==0.9.6
xlwt==1.3.0

